refactor(test): group channel lifecycle coverage
