+++ test/cli-coverage-sequencer.test.ts
@@ -117,9 +117,9 @@ describe("stable CLI coverage sharding", () => {
     expect(Object.fromEntries(owners)).toEqual({
       "cli:src/lib/example.test.ts": 6,
       "e2e-support:test/e2e/support/example.test.ts": 8,
-      "integration:test/hermes-restart-config-seal-write-lock.test.ts": 4,
-      "integration:test/local-credential-helper-fields.test.ts": 7,
-      "integration:test/regular-0.test.ts": 4,
+      "integration:test/hermes-restart-config-seal-write-lock.test.ts": 1,
+      "integration:test/local-credential-helper-fields.test.ts": 2,
+      "integration:test/regular-0.test.ts": 8,
     });
   });
